requirements.txt
pymycobot>=3.6.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
openai>=1.12.0
pyyaml>=6.0
mcp[cli]>=1.0.0
flask>=3.0.0
httpx>=0.27.0
